librbd: moved object cacher writeback handler to cache sub-directory
